+++ src/main/java/io/supertokens/webserver/api/session/SessionRegenerateAPI.java
@@ -44,7 +44,6 @@ import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 import java.security.SignatureException;
 import java.security.spec.InvalidKeySpecException;
-import java.util.Arrays;
 
 public class SessionRegenerateAPI extends WebserverAPI {