+++ src/main/java/io/supertokens/webserver/api/session/SessionRemoveAPI.java
@@ -51,7 +51,7 @@ public class SessionRemoveAPI extends WebserverAPI {
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
-        // API is tenant specific
+        // API is tenant specific. also operates on all tenants in an app if `revokeAcrossAllTenants` is set to true
         JsonObject input = InputParser.parseJsonObjectOrThrowError(req);
 
         String userId = InputParser.parseStringOrThrowError(input, "userId", true);
@@ -78,10 +78,26 @@ public class SessionRemoveAPI extends WebserverAPI {
                     new BadRequestException("Invalid JSON input - use one of userId or sessionHandles array"));
         }
 
+        Boolean revokeAcrossAllTenants = InputParser.parseBooleanOrThrowError(input, "revokeAcrossAllTenants", true);
+        if (userId == null && revokeAcrossAllTenants != null) {
+            throw new ServletException(new BadRequestException("Invalid JSON input - revokeAcrossAllTenants can only be set if userId is set"));
+        }
+
+        if (revokeAcrossAllTenants == null) {
+            revokeAcrossAllTenants = true;
+        }
+
         if (userId != null) {
             try {
-                String[] sessionHandlesRevoked = Session.revokeAllSessionsForUser(
-                        this.getTenantIdentifierWithStorageFromRequest(req), userId);
+                String[] sessionHandlesRevoked;
+                if (revokeAcrossAllTenants) {
+                    sessionHandlesRevoked = Session.revokeAllSessionsForUser(
+                            main, this.getAppIdentifierWithStorage(req), userId);
+                } else {
+                    sessionHandlesRevoked = Session.revokeAllSessionsForUser(
+                            main, this.getTenantIdentifierWithStorageFromRequest(req), userId);
+                }
+
                 if (StorageLayer.getStorage(this.getTenantIdentifierWithStorageFromRequest(req), main).getType() ==
                         STORAGE_TYPE.SQL) {
                     try {
@@ -110,8 +126,8 @@ public class SessionRemoveAPI extends WebserverAPI {
             }
         } else {
             try {
-                String[] sessionHandlesRevoked = Session.revokeSessionUsingSessionHandles(
-                        this.getTenantIdentifierWithStorageFromRequest(req), sessionHandles);
+                String[] sessionHandlesRevoked = Session.revokeSessionUsingSessionHandles(main,
+                        this.getAppIdentifierWithStorage(req), sessionHandles);
                 JsonObject result = new JsonObject();
                 result.addProperty("status", "OK");
                 JsonArray sessionHandlesRevokedJSON = new JsonArray();